+++ CHANGELOG.md
@@ -9,6 +9,7 @@
         - Don't name decks after patterns :(
     - You can set an intensity value in the `decks.ini` file
     - You can leave holes in the deck with `_`
+    - Pattern names now display properly
 - "Soft snap" for MIDI controllers
     - If you modify a slider with mouse/keyboard, you need to move the MIDI knob near it before it can be controlled again.
     - (Just try it.)